splits add-provider into node and rpcurl flavors
