+++ hyperdrive/packages/terminal/Cargo.toml
@@ -1,7 +1,8 @@
 [workspace]
 resolver = "2"
 members = [
-    "add-provider",
+    "add-node-provider",
+    "add-rpcurl-provider",
     "alias",
     "cat",
     "echo",
