+++ hyperdrive/packages/terminal/add-node-provider/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "add-provider"
+name = "add-node-provider"
 version = "0.1.0"
 edition = "2021"
