+++ hyperdrive/packages/terminal/add-rpcurl-provider/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "add-rpcurl-provider"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+simulation-mode = []
+
+[dependencies]
+hyperware_process_lib = "2.0.0"
+rmp-serde = "1.1.2"
+serde = { version = "1.0", features = ["derive"] }
+wit-bindgen = "0.42.1"
+serde_json = "1.0.140"
+
+[lib]
+crate-type = ["cdylib"]
+
+[package.metadata.component]
+package = "hyperware:process"
